stable ish redux components for auth and fireStore
