+++ client/app/components/accountEmailPassword.js
@@ -1,7 +1,9 @@
 import React, {useState} from 'react';
 import { getFirebaseAuth, getFirebaseFirestore } from '../lib/firebaseClient';
-import { doc, setDoc } from 'firebase/firestore';
-import { createUserWithEmailAndPassword, signInWithEmailAndPassword } from 'firebase/auth';
+// import { doc, setDoc } from 'firebase/firestore';
+import { createUserWithEmailAndPassword, signInWithEmailAndPassword, updateCurrentUser } from 'firebase/auth';
+import { createUserDocument, updateUserStatus } from '../redux/authSlice';
+import { useDispatch } from 'react-redux';
 import { useRouter } from 'next/navigation';
 import styles from "../page.module.css";
 import Image from "next/image";
@@ -11,26 +13,12 @@ export default function AccountEmailPassword({action}) {
     const [email, setEmail] = useState("");
     const [password, setPassword] = useState("");
     const db = getFirebaseFirestore();
+    const dispatch = useDispatch();
 
     const router = useRouter();
     const checkPasswordRequirements = (password) => {
         const passwordRegex = /^(?=.*\d)(?=.*[a-z])(?=.*[A-Z])(?=.*[a-zA-Z]).{8,}$/;
         return passwordRegex.test(password);
-    }
-
-    const createUserDocument = async(user) => {
-        try{
-            const userRef = doc(db, 'users', user.uid);
-            await setDoc(userRef, {
-                email: user.email,
-                uid: user.uid,
-                displayName: user.email,
-                status: 'online',
-                lastLogin: new Date(),
-            });
-        } catch (error) {
-            console.error('Error creating user document:', error);
-        }
     }
     
     const createNewAccount = async() => {
@@ -61,7 +49,7 @@ export default function AccountEmailPassword({action}) {
                 });
                 const data = await response.json();
                 if (response.ok) {
-                    await createUserDocument(user);
+                    await dispatch(createUserDocument(user));
                     alert("Account created!");
                     router.push(`/${uid}`);
                 }else if(response.status === 400) {
@@ -104,7 +92,7 @@ export default function AccountEmailPassword({action}) {
             });
             const data = await response.json();
             if (response.ok) {
-                await createUserDocument(user);
+                await dispatch(updateUserStatus(user.uid, 'online'));
                 alert("Signed in!");
                 router.push(`/${uid}`);
             }else if(response.status === 400) {
