+++ client/app/redux/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit'
 import dataReducer from './data'
-import authReducer from '../authSlice'
+import authReducer from './authSlice'
 export const store = configureStore({
   reducer: {
     selectedData: dataReducer,
